split file by email category
