+++ app/services/emails/EmailDefinitionsVarious.scala
@@ -1,42 +1,19 @@
 package services.emails
 
-import enumeratum.EnumEntry
-import enumeratum.PlayEnum
 import models.User
 import models.auth.AuthToken
-import EmailsExamplesUtils._
 import services.emails.EmailCategory.Various
+import services.emails.EmailsExamplesUtils._
 import utils.EmailAddress
 import utils.EmailSubjects
 import utils.FrontRoute
 
 import java.net.URI
-
-sealed trait EmailCategory extends EnumEntry
-
-object EmailCategory extends PlayEnum[EmailCategory] {
-  override def values: IndexedSeq[EmailCategory] = findValues
-
-  case object Various extends EmailCategory
-
-  case object Admin extends EmailCategory
-
-  case object Dgccrf extends EmailCategory
-  case object Pro    extends EmailCategory
-  case object Conso  extends EmailCategory
-
-}
-
-trait EmailDefinition {
-  val category: EmailCategory
-  def examples: Seq[(String, EmailAddress => Email)]
-
-}
-
-object EmailDefinitions {
+object EmailDefinitionsVarious {
 
   case object ResetPassword extends EmailDefinition {
     override val category = Various
+
     override def examples =
       Seq("reset_password" -> (recipient => build(genUser.copy(email = recipient), genAuthToken)))
 
@@ -44,6 +21,7 @@ object EmailDefinitions {
       new Email {
         override val recipients: List[EmailAddress] = List(user.email)
         override val subject: String                = EmailSubjects.RESET_PASSWORD
+
         override def getBody: (FrontRoute, EmailAddress) => String = (frontRoute, contactAddress) =>
           views.html.mails.resetPassword(user, authToken)(frontRoute, contactAddress).toString
       }
@@ -59,9 +37,9 @@ object EmailDefinitions {
       new Email {
         override val recipients: List[EmailAddress] = List(recipient)
         override val subject: String                = EmailSubjects.UPDATE_EMAIL_ADDRESS
+
         override def getBody: (FrontRoute, EmailAddress) => String = (_, _) =>
           views.html.mails.updateEmailAddress(invitationUrl, daysBeforeExpiry).toString()
       }
   }
-
 }
